+++ django-data/image/biosample/tasks/__init__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 16 11:26:52 2019
+
+@author: Paolo Cozzi <cozzi@ibba.cnr.it>
+"""
+
+from .fetch import FetchStatusTask
+from .submit import SubmitTask
+
+__all__ = ["SubmitTask", "FetchStatusTask"]
